Added pre-compiled headers and Jupiter primitive data types aliases.

+++ jupiter_engine/src/events/EventHandler.h
@@ -18,7 +18,7 @@ public:
 
     virtual void execute(const Event& event) = 0;
 
-    virtual size_t getHashCode() const = 0;
+    virtual uint64 getHashCode() const = 0;
 };
 
 template <typename EventT>
